+++ nvim/.config/nvim/after/plugin/colorscheme.lua
@@ -3,29 +3,29 @@
 -- vim.cmd [[highlight ColorColumn ctermbg=0 guibg=darkgrey]]
 -- vim.cmd [[highlight CursorLine ctermbg=0 guibg=#262626]]
 
-require("gruvbox").setup({
-  undercurl = true,
-  underline = true,
-  bold = true,
-  italic = true,
-  strikethrough = true,
-  invert_selection = false,
-  invert_signs = false,
-  invert_tabline = false,
-  invert_intend_guides = false,
-  inverse = true, -- invert background for search, diffs, statuslines and errors
-  contrast = "hard", -- can be "hard", "soft" or empty string
-  palette_overrides = {},
-  overrides = {},
-  dim_inactive = false,
-  transparent_mode = false,
-})
+-- require("gruvbox").setup({
+--   undercurl = true,
+--   underline = true,
+--   bold = true,
+--   italic = true,
+--   strikethrough = true,
+--   invert_selection = false,
+--   invert_signs = false,
+--   invert_tabline = false,
+--   invert_intend_guides = false,
+--   inverse = true, -- invert background for search, diffs, statuslines and errors
+--   contrast = "hard", -- can be "hard", "soft" or empty string
+--   palette_overrides = {},
+--   overrides = {},
+--   dim_inactive = false,
+--   transparent_mode = false,
+-- })
 
-vim.g.gruvbox_invert_selection='0'
-vim.g.gruvbox_italic = 1
-vim.g.gruvbox_contrast_dark = 'hard'
-vim.cmd [[set background=dark]]
-vim.cmd [[colorscheme gruvbox]]
+-- vim.g.gruvbox_invert_selection='0'
+-- vim.g.gruvbox_italic = 1
+-- vim.g.gruvbox_contrast_dark = 'hard'
+-- vim.cmd [[set background=dark]]
+-- vim.cmd [[colorscheme gruvbox]]
 
 -- vim.g.falcon_background = false
 -- vim.g.falcon_inactive = true
@@ -53,7 +53,7 @@ vim.cmd [[colorscheme gruvbox]]
 -- vim.g.tokyonight_transparent = true
 -- vim.cmd [[colorscheme tokyonight]]
 
-vim.cmd [[highlight Normal guibg=NONE ctermbg=NONE]]
+-- vim.cmd [[highlight Normal guibg=NONE ctermbg=NONE]]
 -- vim.cmd [[highlight ColorColumn ctermbg=0 guibg=darkgrey]]
 -- vim.cmd [[highlight CursorLine ctermbg=0 guibg=#262626]]
 -- vim.cmd [[highlight CursorLine ctermbg=0 guibg=#555555]]
