--- nvim/.config/nvim/after/plugin/init.lua
@@ -1,1 +1,0 @@
--- print("hello from after init.lua")
