--- nvim/.config/nvim/after/plugin/netrw.lua
@@ -1,5 +1,0 @@
-vim.g.netrw_browse_split = 2
-vim.g.netrw_banner=0        -- disable annoying banner
-vim.g.netrw_browse_split=4  -- open in prior window
-vim.g.netrw_altv=1          -- open splits to the right
-vim.cmd [[let g:netrw_list_hide=netrw_gitignore#Hide()]]
