+++ nvim/.config/nvim/lua/plugins/init.lua
@@ -1,35 +1,10 @@
-print("this is in plugsin/init.lua")
-
 return {
-		{
-			'junegunn/fzf',
-			build = ':call fzf#install()'
-		},
-		{
-			'junegunn/fzf.vim',
-			init = function()
-				vim.env.FZF_DEFAULT_COMMAND = 'rg --files --hidden'
-				vim.g.fzf_layout = {
-					up ='~90%',
-					window = {
-						width = 0.8,
-						height = 0.8,
-						yoffset = 0.5,
-						xoffset = 0.5,
-						highlight = 'Todo',
-						border = 'sharp'
-					}
-				}
-			end
-		},
-		'lewis6991/gitsigns.nvim',
 		{
 			'dstein64/vim-startuptime',
 			cmd = 'StartupTime'
 		},
+		'dyng/ctrlsf.vim',
 		{'christoomey/vim-tmux-navigator', lazy = true},
-		{'dyng/ctrlsf.vim', lazy = true},
-		{'jremmen/vim-ripgrep', lazy = true},
 		{'mbbill/undotree', lazy = true},
 		{'editorconfig/editorconfig-vim', lazy = true},
 		{'jpalardy/vim-slime', branch = 'main'},
@@ -45,19 +20,6 @@ return {
 		{'cespare/vim-toml', ft = 'toml'},
 		{'elixir-editors/vim-elixir', ft = { 'elixir', 'eelixir' }},
 		{'udalov/kotlin-vim', ft = 'kotlin'},
-		{
-			'nvim-lualine/lualine.nvim',
-			dependencies = {
-				'kyazdani42/nvim-web-devicons',
-			}
-		},
-
-		-- colorschemes
-		'ellisonleao/gruvbox.nvim',
-		{'fenetikm/falcon', lazy = true},
-		{'NLKNguyen/papercolor-theme', lazy = true},
-		{'ishan9299/nvim-solarized-lua', lazy = true},
-		{'folke/tokyonight.nvim', lazy = true},
 
 		-- Tim Pope
 		'tpope/vim-fugitive',
@@ -65,62 +27,14 @@ return {
 		'tpope/vim-surround',
 		'tpope/vim-repeat',
 		'tpope/vim-commentary',
-		{'tpope/vim-unimpaired', lazy = true},
+		'tpope/vim-unimpaired',
 
-		'ThePrimeagen/harpoon',
-		'ThePrimeagen/git-worktree.nvim',
 		'ThePrimeagen/vim-be-good',
 
-		-- Tree sitter
 		{
-			'nvim-treesitter/nvim-treesitter',
-			build = ':TSUpdate',
+			'nvim-lua/popup.nvim',
 			dependencies = {
-				'nvim-treesitter/playground',
-			}
-		},
-
-		-- ' LSP stuff
-		'neovim/nvim-lspconfig',
-		'nvim-lua/popup.nvim',
-		{
-			'j-hui/fidget.nvim',
-			lazy = true
-		},
-
-		-- Telecope
-		{
-			'nvim-telescope/telescope.nvim',
-			dependencies = {
-				'nvim-lua/plenary.nvim',
-				'nvim-telescope/telescope-live-grep-args.nvim',
-				{
-					'nvim-telescope/telescope-fzf-native.nvim',
-					build = 'make',
-					branch = 'main',
-					dependencies = {
-						'nvim-telescope/telescope.nvim',
-					}
-				},
-			}
-		},
-
-		-- Snippets
-		{
-			'L3MON4D3/LuaSnip',
-			config = function()
-				vim.g.snippets = 'luasnip'
-			end,
-			dependencies = {
-				'saadparwaiz1/cmp_luasnip',
-				'rafamadriz/friendly-snippets',
-			}
-		},
-
-		-- completion
-		'hrsh7th/nvim-cmp',
-		'hrsh7th/cmp-buffer',
-		'hrsh7th/cmp-nvim-lua',
-		'hrsh7th/cmp-nvim-lsp',
-		'onsails/lspkind-nvim'
+				'nvim-lua/plenary.nvim'
+			},
+		}
 	}
