+++ nvim/.config/nvim/lua/plugins/lspkind.lua
@@ -1,0 +1,10 @@
+return {
+	{
+		'onsails/lspkind-nvim',
+		config = function()
+			require('lspkind').init({
+							mode = 'symbol_text',
+			})
+		end
+	}
+}
